+++ cpp/libjoynr/joynr-messaging/AbstractJoynrMessagingConnector.cpp
@@ -32,18 +32,14 @@ AbstractJoynrMessagingConnector::AbstractJoynrMessagingConnector(
         const std::string& interfaceName,
         const std::string& proxyParticipantId,
         const std::string& providerParticipantId,
-        const MessagingQos& qosSettings,
-        IClientCache* cache,
-        bool cached)
+        const MessagingQos& qosSettings)
         : joynrMessageSender(joynrMessageSender),
           subscriptionManager(subscriptionManager),
           domain(domain),
           interfaceName(interfaceName),
           proxyParticipantId(proxyParticipantId),
           providerParticipantId(providerParticipantId),
-          qosSettings(qosSettings),
-          cache(cache),
-          cached(cached)
+          qosSettings(qosSettings)
 {
 }