+++ cpp/tests/utils/MockObjects.h
@@ -48,7 +48,6 @@
 #include "joynr/IMessageReceiver.h"
 #include "joynr/IDispatcher.h"
 #include "joynr/IMessaging.h"
-#include "joynr/IClientCache.h"
 #include "joynr/ReplyCaller.h"
 #include "joynr/ISubscriptionListener.h"
 #include "joynr/MessagingQos.h"
@@ -549,12 +548,6 @@ public:
     }
 };
 
-class MockClientCache : public joynr::IClientCache {
-public:
-   MOCK_METHOD1(lookUp, boost::any(const std::string& attributeId));
-   MOCK_METHOD2(insert, void(std::string attributeId, boost::any value));
-};
-
 class MockDiscovery : public joynr::system::IDiscovery {
 public:
     MOCK_METHOD1(
@@ -695,37 +688,27 @@ public:
         RoutingProxy(
                 std::make_shared<const joynr::system::RoutingTypes::Address>(),
                 nullptr,
-                nullptr,
                 "domain",
-                joynr::MessagingQos(),
-                false),
+                joynr::MessagingQos()),
         ProxyBase(
                 nullptr,
-                nullptr,
                 "domain",
-                joynr::MessagingQos(),
-                false),
+                joynr::MessagingQos()),
         RoutingProxyBase(
                 std::make_shared<const joynr::system::RoutingTypes::Address>(),
                 nullptr,
-                nullptr,
                 "domain",
-                joynr::MessagingQos(),
-                false),
+                joynr::MessagingQos()),
         RoutingSyncProxy(
                 std::make_shared<const joynr::system::RoutingTypes::Address>(),
                 nullptr,
-                nullptr,
                 "domain",
-                joynr::MessagingQos(),
-                false),
+                joynr::MessagingQos()),
         RoutingAsyncProxy(
                 std::make_shared<const joynr::system::RoutingTypes::Address>(),
                 nullptr,
-                nullptr,
                 "domain",
-                joynr::MessagingQos(),
-                false)
+                joynr::MessagingQos())
     { }
 
     MOCK_METHOD3(resolveNextHopAsync,
@@ -1038,37 +1021,27 @@ public:
         GlobalDomainAccessControllerProxy(
                 std::make_shared<const joynr::system::RoutingTypes::Address>(),
                 nullptr,
-                nullptr,
-                "domain",
-                joynr::MessagingQos(),
-                false),
+                  "domain",
+                joynr::MessagingQos()),
         ProxyBase(
                 nullptr,
-                nullptr,
                 "domain",
-                joynr::MessagingQos(),
-                false),
+                joynr::MessagingQos()),
         GlobalDomainAccessControllerProxyBase(
                 std::make_shared<const joynr::system::RoutingTypes::Address>(),
                 nullptr,
-                nullptr,
                 "domain",
-                joynr::MessagingQos(),
-                false),
+                joynr::MessagingQos()),
         GlobalDomainAccessControllerSyncProxy(
                 std::make_shared<const joynr::system::RoutingTypes::Address>(),
                 nullptr,
-                nullptr,
                 "domain",
-                joynr::MessagingQos(),
-                false),
+                joynr::MessagingQos()),
         GlobalDomainAccessControllerAsyncProxy(
                 std::make_shared<const joynr::system::RoutingTypes::Address>(),
                 nullptr,
-                nullptr,
                 "domain",
-                joynr::MessagingQos(),
-                false)
+                joynr::MessagingQos())
     {
     }